solucionado liberarRecursos y problema con cantidad de entrenadores en deteccionDeadlock

+++ mapas/nivel-test.h
@@ -68,7 +68,7 @@ typedef struct {
 	int batalla;
 	char* algoritmo;
 	int quantum;
-	long retardo;
+	int retardo;
 	char* ip;
 	char* puerto;
 }metadata;
